requirement.txt
django
django-bootstrap-v5
pillow